Move preference extension functions to utils
