+++ app/src/main/java/eu/kanade/tachiyomi/widget/TachiyomiBottomNavigationView.kt
@@ -20,7 +20,7 @@ import androidx.lifecycle.lifecycleScope
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import eu.kanade.tachiyomi.R
 import eu.kanade.tachiyomi.data.preference.PreferencesHelper
-import eu.kanade.tachiyomi.data.preference.asImmediateFlow
+import eu.kanade.tachiyomi.util.preference.asImmediateFlow
 import eu.kanade.tachiyomi.util.system.applySystemAnimatorScale
 import kotlinx.coroutines.flow.launchIn
 import uy.kohesive.injekt.Injekt
